Add mem streams and DOS90 compatibility
This commit introduces a POSIX-style memory stream layer for SMALLC and adds stream flag helpers for mode parsing. It also reorganizes the BIOS/DOS headers into BIOS90/DOS90 compatibility directories, adds and updates DOS error and file attribute definitions, and narrows the CMake source glob to SMALLC to avoid building legacy DOS/BIOS sources. A binary TEST.exe was also added to the repo.

+++ src/DOS/dos_error_types.h
@@ -6,6 +6,6 @@
 #ifndef DOS_ERROR_TYPES_H
 #define DOS_ERROR_TYPES_H
 
-typedef unsigned short dos_error_code_t;
+typedef int dos_error_code_t;
 
 #endif
